+++ src/common/irrelevant-entities-container.ts
@@ -1,0 +1,11 @@
+export class IrrelevantEntitiesContainer {
+    irrelevantContainer: { [key: string]: string[] };
+
+    constructor() {
+        this.irrelevantContainer = {};
+    }
+
+    addIrrelevantEntitiesOfQuery(queryKey: string, irrelevantEntities: string[]): void {
+        this.irrelevantContainer[queryKey] = irrelevantEntities;
+    }
+}