+++ src/http/request/polaris-request-headers.ts
@@ -1,3 +1,5 @@
+import { PolarisRequestHeaders } from '@enigmatis/utills';
+import { UserInputError } from 'apollo-server-koa';
 import * as joi from 'joi';
 
 const headersSchema = joi.object().keys({
@@ -13,25 +15,12 @@ const headersSchema = joi.object().keys({
     'requesting-sys-name': joi.string(),
 });
 
-export interface PolarisRequestHeaders {
-    dataVersion?: number;
-    isSnapshot?: boolean;
-    includeLinkedOperation?: boolean;
-    snapshotPageSize?: number;
-    requestId?: string;
-    upn?: string;
-    eventKind?: string;
-    realityId?: number;
-    requestingSystemId?: string;
-    requestingSystemName?: string;
-}
-
 export const getHeaders = (candidate: object): PolarisRequestHeaders => {
     const { error, value: validatedHeaders } = joi.validate(candidate, headersSchema, {
         stripUnknown: true,
     }) as { error: joi.ValidationError; value: { [key: string]: any } };
     if (error) {
-        throw error;
+        throw new UserInputError(error.message);
     } else {
         return {
             dataVersion: validatedHeaders['data-version'],