+++ src/inversion-of-control/polaris-types.ts
@@ -3,13 +3,13 @@ import {
     MiddlewaresConfig,
     PolarisServerConfig,
 } from '../common/injectable-interfaces';
-import { PolarisGraphqlLogger } from '../logging/polaris-graphql-logger';
+import { PolarisGraphQLLogger } from '../logging/polaris-graphql-logger';
 import { GraphQLServer } from '../server/graphql-server';
 
 export const POLARIS_TYPES = {
     GraphQLServer: Symbol('GraphQLServer'),
     GraphQLSchema: Symbol('GraphQLSchema'),
-    GraphqlLogger: Symbol('PolarisGraphqlLogger'),
+    GraphQLLogger: Symbol('PolarisGraphQLLogger'),
     PolarisHeadersFactory: Symbol('PolarisHeadersFactory'),
     Middleware: Symbol('Middleware'),
     PolarisServerConfig: Symbol('PolarisServerConfig'),