+++ src/main.ts
@@ -11,9 +11,14 @@ export {
 } from './common/injectable-interfaces';
 export { polarisContainer } from './inversion-of-control/container-manager';
 export { POLARIS_TYPES } from './inversion-of-control/polaris-types';
-export { PolarisGraphqlLogger } from './logging/polaris-graphql-logger';
+export { PolarisGraphQLLogger } from './logging/polaris-graphql-logger';
 export { PolarisProperties } from './properties/polaris-properties';
 export { RepositoryEntity } from './dal/entities/repository-entity';
+export { CommonEntityInterface } from './schema/common/common-entity-interface';
 export { GraphQLServer, PolarisGraphQLServer } from './server/graphql-server';
 export { MiddlewaresConfiguration } from './middlewares/middleware';
-export { makeExecutablePolarisSchema } from './schema/polaris-schema-creator';
+export { PolarisContext } from './server/polaris-context';
+export {
+    buildPolarisSchema,
+    makeExecutablePolarisSchema,
+} from './schema/utils/polaris-schema-creator';