+++ src/server/graphql-server.ts
@@ -1,87 +1,145 @@
-import { PolarisLogger } from '@enigmatis/polaris-logs';
-import { ApolloServer, Config } from 'apollo-server-koa';
-import { GraphQLSchema } from 'graphql';
+import { ApolloServer, Config, PubSub } from 'apollo-server-koa';
+import { GraphQLError, GraphQLFormattedError, GraphQLSchema } from 'graphql';
 import { applyMiddleware } from 'graphql-middleware';
+import * as http from 'http';
 import { inject, injectable, multiInject } from 'inversify';
 import * as Koa from 'koa';
 import * as koaBody from 'koa-bodyparser';
 import { PolarisServerConfig } from '../common/injectable-interfaces';
+import { IrrelevantEntitiesContainer } from '../common/irrelevant-entities-container';
 import { getHeaders } from '../http/request/polaris-request-headers';
 import { POLARIS_TYPES } from '../inversion-of-control/polaris-types';
+import { PolarisGraphQLLogger } from '../logging/polaris-graphql-logger';
 import { Middleware } from '../middlewares/middleware';
 import { createMiddleware } from '../middlewares/polaris-middleware-creator';
 import { PolarisProperties } from '../properties/polaris-properties';
+import { IrrelevantEntitiesExtension } from './irrelevant-entities-extension';
 import { PolarisContext } from './polaris-context';
 
-const app = new Koa();
-app.use(koaBody());
-
 export interface GraphQLServer {
-    start(): void;
+    start(app?: Koa<any, {}>): void;
+    stop(): void;
 }
+
+export type contextCreator = (ctx: Koa.Context) => object;
 
 @injectable()
 export class PolarisGraphQLServer implements GraphQLServer {
-    @inject(POLARIS_TYPES.GraphqlLogger) polarisLogger!: PolarisLogger;
+    private app: Koa;
     private server: ApolloServer;
     private polarisProperties: PolarisProperties;
+    private customContexts: contextCreator[] = [];
+    private httpServer?: http.Server;
 
     constructor(
         @inject(POLARIS_TYPES.GraphQLSchema) schema: GraphQLSchema,
         @inject(POLARIS_TYPES.PolarisServerConfig) propertiesConfig: PolarisServerConfig,
         @multiInject(POLARIS_TYPES.Middleware) middlewares: Middleware[],
+        @inject(POLARIS_TYPES.GraphQLLogger) private polarisLogger: PolarisGraphQLLogger,
     ) {
-        const executableSchemaWithMiddlewares = applyMiddleware(
+        const executableSchemaWithMiddleware = applyMiddleware(
             schema,
             ...(middlewares.map(createMiddleware) as any),
         );
         this.polarisProperties = propertiesConfig.polarisProperties;
-        const config: Config = {
-            schema: executableSchemaWithMiddlewares,
-            context: ({ ctx }: { ctx: Koa.Context }): PolarisContext => {
-                try {
-                    const headers = getHeaders(ctx.request.headers);
-                    return {
-                        headers,
-                        body: ctx.request.body,
-                    };
-                } catch (e) {
-                    this.polarisLogger.error('Headers error', { throwable: e });
-                    throw new Error('Unable to format headers');
-                }
-            },
-            formatError: (error: any) => {
-                if (error.extensions && error.extensions.code === 'GRAPHQL_VALIDATION_FAILED') {
-                    return error;
-                } else {
-                    this.polarisLogger.error('Apollo server error', { throwable: error });
-                    return new Error('Internal server error');
-                }
-            },
 
-            formatResponse: (response: any) => {
-                if (!response.data.__schema && !response.data.__type) {
-                    this.polarisLogger.info(
-                        `Finished response, answer is ${JSON.stringify(response)}`,
-                    );
-                }
-                return response;
-            },
+        const config: Config = {
+            schema: executableSchemaWithMiddleware,
+            context: (args: { ctx: Koa.Context; connection: any }) => this.getContext(args),
+            formatError: (error: any) => this.formatError(error),
+            formatResponse: (response: any) => this.formatResponse(response),
+            extensions: [() => new IrrelevantEntitiesExtension()],
         };
         this.server = new ApolloServer(config);
-        if (this.polarisProperties.endpoint) {
-            this.server.applyMiddleware({ app, path: this.polarisProperties.endpoint });
-        } else {
-            this.server.applyMiddleware({ app });
+        this.app = new Koa();
+        this.app.use(koaBody());
+    }
+
+    async start(app = this.app) {
+        return new Promise(resolve => {
+            if (this.polarisProperties.endpoint) {
+                this.server.applyMiddleware({ app, path: this.polarisProperties.endpoint });
+            } else {
+                this.server.applyMiddleware({ app });
+            }
+            const port = this.polarisProperties.port;
+            this.httpServer = app.listen(port, () => {
+                this.polarisLogger.info(
+                    `🚀 Server ready at http://localhost:${port}${this.server.graphqlPath}`,
+                );
+                this.polarisLogger.info(
+                    `🚀 Subscriptions ready at ws://localhost:${port}${
+                        this.server.subscriptionsPath
+                    }`,
+                );
+                resolve();
+            });
+            if (this.polarisProperties.includeSubscription) {
+                this.server.installSubscriptionHandlers(this.httpServer);
+            }
+        });
+    }
+
+    stop() {
+        if (this.httpServer) {
+            this.httpServer.close();
         }
     }
 
-    start() {
-        const port = this.polarisProperties.port;
-        app.listen(port, () => {
-            this.polarisLogger.info(
-                `🚀 Server ready at http://localhost:${port}${this.server.graphqlPath}`,
-            );
+    addContextCreator(...contextCreators: contextCreator[]) {
+        this.customContexts.push(...contextCreators);
+    }
+
+    private formatResponse(response: any) {
+        if (response.data && !response.data.__schema && !response.data.__type) {
+            this.polarisLogger.info(`Finished response, answer is ${JSON.stringify(response)}`);
+        }
+        return response;
+    }
+
+    private formatError(error: GraphQLError): GraphQLFormattedError {
+        this.polarisLogger.error('Apollo server error', {
+            polarisLogProperties: { throwable: error },
         });
+
+        return {
+            message: error.message,
+            path: error.path,
+            locations: error.locations,
+            extensions: error.extensions,
+        };
+    }
+
+    private getContext({ ctx, connection }: { ctx: Koa.Context; connection: any }): PolarisContext {
+        try {
+            if (!connection) {
+                return this.getHttpContext(ctx);
+            } else {
+                return { body: {}, pubSub: new PubSub() } as any;
+            }
+        } catch (e) {
+            this.polarisLogger.error('Headers error', {
+                polarisLogProperties: { throwable: e },
+            });
+            throw new Error('Unable to format headers');
+        }
+    }
+
+    private getHttpContext(ctx: Koa.Context) {
+        const headers = getHeaders(ctx.request.headers);
+        const context: PolarisContext = {
+            headers,
+            body: ctx.request.body,
+            irrelevantEntities: new IrrelevantEntitiesContainer(),
+            ...this.getCustomContext(ctx),
+        };
+        if (this.polarisProperties.includeSubscription) {
+            context.pubSub = new PubSub();
+        }
+        return context;
+    }
+
+    private getCustomContext(ctx: Koa.Context): object[] {
+        return this.customContexts.map(creator => creator(ctx));
     }
 }