+++ src/server/polaris-context.ts
@@ -1,6 +1,10 @@
-import { PolarisRequestHeaders } from '../http/request/polaris-request-headers';
+import { PolarisBaseContext, PolarisRequestHeaders } from '@enigmatis/utills';
+import { PubSub } from 'apollo-server-koa';
+import { IrrelevantEntitiesContainer } from '../common/irrelevant-entities-container';
 
-export interface PolarisContext {
+export interface PolarisContext extends PolarisBaseContext {
     headers: PolarisRequestHeaders;
     body: any;
+    pubSub?: PubSub;
+    irrelevantEntities: IrrelevantEntitiesContainer;
 }