+++ test/integration-tests/test-server/schema/schema.ts
@@ -1,0 +1,12 @@
+import { GraphQLSchema } from 'graphql';
+import { makeExecutablePolarisSchema } from '../../../../src/main';
+import { resolvers } from './resolvers/book-resolvers';
+import { Book, BookInput, Mutation, Query, Subscription } from './types/schema-types';
+
+export const schema: GraphQLSchema = makeExecutablePolarisSchema({
+    typeDefs: [Book, BookInput, Mutation, Query, Subscription],
+    resolvers: [resolvers],
+    resolverValidationOptions: {
+        requireResolversForResolveType: false,
+    },
+});