+++ test/integration-tests/test-server/schema/types/mutation/root-mutation.ts
@@ -1,0 +1,5 @@
+export const Mutation = `type Mutation {
+                            createBook(book: BookInput!): Book,
+                            updateBook(bookId: String!, update: UpdateBookInput): Book,
+                            deleteBook(bookId: String!): Book,
+                         }`;