+++ test/integration-tests/test-server/schema/types/query/root-query.ts
@@ -1,0 +1,6 @@
+export const Query = `type Query {
+                        books: [Book]
+                        bookById(bookId:String!): Book
+                        deletedBookById(bookId:String!): Book
+                        booksStartsWith(startsWith:String!): [Book]
+                      }`;