+++ test/integration-tests/test-server/schema/types/schema-types.ts
@@ -1,0 +1,5 @@
+export { BookInput } from './input/book-input';
+export { Mutation } from './mutation/root-mutation';
+export { Book } from './output/book';
+export { Query } from './query/root-query';
+export { Subscription } from './subscription/root-subscription';