+++ test/middlewares/middleware-activation-condition/filter-soft-delete.test.ts
@@ -1,11 +1,16 @@
 import { GraphQLResolveInfo } from 'graphql';
+import { IrrelevantEntitiesContainer } from '../../../src/common/irrelevant-entities-container';
 import { ResponseMiddlewareParams } from '../../../src/middlewares/middleware';
 import { SoftDeleteFilter } from '../../../src/middlewares/middleware-activation-condition/filter-soft-delete';
 import { PolarisContext } from '../../../src/server/polaris-context';
 
 describe('filter-soft-delete tests', () => {
     const args = {};
-    const context: PolarisContext = { headers: {}, body: {} };
+    const context: PolarisContext = {
+        headers: {},
+        body: {},
+        irrelevantEntities: new IrrelevantEntitiesContainer(),
+    };
     const info: { [T in keyof GraphQLResolveInfo]: any } = {} as any;
     const root = undefined;
 
@@ -17,7 +22,7 @@ describe('filter-soft-delete tests', () => {
             info,
             result: { deleted: true },
         };
-        expect(SoftDeleteFilter.shouldBeReturned(middlewareParams)).toBe(false);
+        expect(SoftDeleteFilter.shouldBeReturned(middlewareParams.result)).toBe(false);
     });
 
     test('not deleted entity should be returned', () => {