+++ test/server/graphql-server.test.ts
@@ -22,52 +22,116 @@ jest.mock('graphql-middleware', () => ({
 jest.mock('../../src/middlewares/polaris-middleware-creator', () => ({
     createMiddleware: jest.fn(),
 }));
-jest.mock('../../src/schema/polaris-schema-creator', () => ({
-    makeExecutablePolarisSchema: jest.fn(),
-}));
 
 describe('graphql-server tests', () => {
-    const schema: { [T in keyof GraphQLSchema]: any } = {} as any;
+    const graphQLSchemaMock: { [T in keyof GraphQLSchema]: any } = {
+        generateSchema: jest.fn(() => ({ def: 'definition', resolvers: jest.fn() })),
+    } as any;
+    const polarisLog = {
+        info: jest.fn(),
+    } as any;
     const polarisServerConfigMock: { [T in keyof PolarisServerConfig]: any } = {
         polarisProperties: jest.fn(),
     } as any;
     const polarisMiddlewareMock: { [T in keyof PolarisMiddleware]: any } = {} as any;
 
     test('creating new polaris server - with arguments - graphql apply middleware have been called', () => {
-        const server = new PolarisGraphQLServer(schema, polarisServerConfigMock, [
-            polarisMiddlewareMock,
-        ]);
+        const server = new PolarisGraphQLServer(
+            graphQLSchemaMock,
+            polarisServerConfigMock,
+            [polarisMiddlewareMock],
+            polarisLog,
+        );
 
         expect(applyMiddleware).toHaveBeenCalled();
     });
 
     test('creating new polaris server - with arguments - create middleware have been called number of times as middlewares provided', () => {
         const middlewares = [polarisMiddlewareMock, polarisMiddlewareMock, polarisMiddlewareMock];
-        const server = new PolarisGraphQLServer(schema, polarisServerConfigMock, middlewares);
+        const server = new PolarisGraphQLServer(
+            graphQLSchemaMock,
+            polarisServerConfigMock,
+            middlewares,
+            polarisLog,
+        );
 
         expect(createMiddleware).toHaveBeenCalledTimes(middlewares.length);
     });
 
     test('creating new polaris server - with arguments - apollo server constructor have been called', () => {
-        const server = new PolarisGraphQLServer(schema, polarisServerConfigMock, [
-            polarisMiddlewareMock,
-        ]);
+        const server = new PolarisGraphQLServer(
+            graphQLSchemaMock,
+            polarisServerConfigMock,
+            [polarisMiddlewareMock],
+            polarisLog,
+        );
 
         expect(ApolloServer).toHaveBeenCalled();
     });
 
-    test('creating new polaris server - with arguments - apollo server apply middleware have been called with custom path', () => {
+    test('on start - apollo server apply middleware have been called with custom path', () => {
         const polarisServerConfigMockWithEndpoint: { [T in keyof PolarisServerConfig]: any } = {
             polarisProperties: { endpoint: 'test' },
         } as any;
 
-        const server = new PolarisGraphQLServer(schema, polarisServerConfigMockWithEndpoint, [
-            polarisMiddlewareMock,
-        ]);
+        const server = new PolarisGraphQLServer(
+            graphQLSchemaMock,
+            polarisServerConfigMockWithEndpoint,
+            [polarisMiddlewareMock],
+            polarisLog,
+        );
+        server.start();
         expect(apolloServerMock.applyMiddleware).toHaveBeenCalledWith(
             expect.objectContaining({
                 path: polarisServerConfigMockWithEndpoint.polarisProperties.endpoint,
             }),
         );
+        server.stop();
+    });
+
+    test('on start - with custom app - apollo server apply middleware have been called with custom app', () => {
+        const polarisServerConfigMockWithEndpoint: { [T in keyof PolarisServerConfig]: any } = {
+            polarisProperties: { endpoint: 'test' },
+        } as any;
+
+        const app = {
+            listen: jest.fn(),
+        } as any;
+
+        const server = new PolarisGraphQLServer(
+            graphQLSchemaMock,
+            polarisServerConfigMockWithEndpoint,
+            [polarisMiddlewareMock],
+            polarisLog,
+        );
+        server.start(app);
+        expect(apolloServerMock.applyMiddleware).toHaveBeenCalledWith(
+            expect.objectContaining({ app }),
+        );
+        expect(app.listen).toHaveBeenCalled();
+        server.stop();
+    });
+
+    test('on start - with custom app - apollo server apply middleware have been called with custom app', () => {
+        const polarisServerConfigMockWithEndpoint: { [T in keyof PolarisServerConfig]: any } = {
+            polarisProperties: { endpoint: 'test' },
+        } as any;
+
+        const app = {
+            listen: jest.fn(),
+        } as any;
+
+        const server = new PolarisGraphQLServer(
+            graphQLSchemaMock,
+            polarisServerConfigMockWithEndpoint,
+            [polarisMiddlewareMock],
+            polarisLog,
+        );
+        server.start(app);
+        expect(apolloServerMock.applyMiddleware).toHaveBeenCalledWith(
+            expect.objectContaining({ app }),
+        );
+        expect(app.listen).toHaveBeenCalled();
+        server.stop();
     });
 });